+++ srcs/ray/hit/hit_cylinder_body.c
@@ -3,21 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   hit_cylinder_body.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mfukui <mfukui@student.42.fr>              +#+  +:+       +#+        */
+/*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/22 16:28:02 by mfukui            #+#    #+#             */
-/*   Updated: 2025/04/22 16:35:00 by mfukui           ###   ########.fr       */
+/*   Updated: 2025/04/22 17:11:49 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-static bool	hit_cylinder_cap_top(t_ray ray, t_cylinder *cy, float *t)
+static bool	hit_cylinder_cap_top(t_ray ray, t_cylinder *cy, double *t)
 {
 	t_vector	axis;
 	t_vector	center;
-	float		denom;
-	float		temp_t;
+	double		denom;
+	double		temp_t;
 	t_vector	hit_point;
 
 	axis = vec_normalize(cy->orientation);
@@ -35,13 +35,13 @@ static bool	hit_cylinder_cap_top(t_ray ray, t_cylinder *cy, float *t)
 	return (true);
 }
 
-static bool	hit_cylinder_cap_bottom(t_ray ray, t_cylinder *cy, float *t)
+static bool	hit_cylinder_cap_bottom(t_ray ray, t_cylinder *cy, double *t)
 {
 	t_vector	axis;
 	t_vector	center;
-	float		denom;
+	double		denom;
 	t_vector	hit_point;
-	float		temp_t;
+	double		temp_t;
 
 	axis = vec_normalize(cy->orientation);
 	center = cy->position;
@@ -59,10 +59,10 @@ static bool	hit_cylinder_cap_bottom(t_ray ray, t_cylinder *cy, float *t)
 	return (true);
 }
 
-bool	hit_cylinder_caps(t_ray ray, t_cylinder *cy, float *t)
+bool	hit_cylinder_caps(t_ray ray, t_cylinder *cy, double *t)
 {
-	float	t_top;
-	float	t_bot;
+	double	t_top;
+	double	t_bot;
 	bool	hit;
 
 	hit = false;
